pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "confluence-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for Atlassian Confluence"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastmcp",
    "requests",
    "beautifulsoup4",
    "chainlit",
    "langgraph",
    "langchain",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-google-genai",
    "python-dotenv",
    "mcp",
    "google-generativeai"
]

[project.scripts]
confluence-mcp = "confluence_mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/confluence_mcp"]
